requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
easyocr==1.7.2
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.2
Keras-Applications==1.0.8
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
ninja==1.11.1.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
Pygments==2.19.1
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
shapely==2.1.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.109
ultralytics-thop==2.0.14
urllib3==2.4.0
validators==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
